Add handling of incorrect input
If the user tries to write a negative value, it will be changed to
zero.

If the user leaves the input field empty, the view model will get
zero

If any of еру "pipeline" settings (speed, distance and spawn delay)
are set to zero, pipeline will not spawn or move cubes.